Move main.py, rename imports
